backend/requirements.txt
amqp==5.2.0
APScheduler==3.10.4
asgiref==3.8.1
billiard==4.2.0
boto3==1.34.98
botocore==1.34.98
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.13
django-cors-headers==4.9.0
djangorestframework==3.16.1
jmespath==1.0.1
kombu==5.3.7
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
s3transfer==0.10.1
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
urllib3==1.26.18
vine==5.1.0
wcwidth==0.2.13
